expanded list of month formats
Added in lower case as well.  I didn't want to just lowercase everything because we need to distinguish between the month of May and the word "may".  Thus, May is the only month with no lower case version.  It will only pick it up if it is capitalized.  Still an issue in the month of March and the word "march".  could implement a ML algorithm to distinguish these at some point.