debugged 24-hour time parsing
Corrected some major flaws in the parsing of 24-hour time.  It was identifying lone times like 1350, but was missing times like 2041EST.  I corrected the has24HourTime() method to correctly return the time string in both instances.  I also corrected the build24HourTime() method as it was calculating the indicies incorrectly.