fixed bugs with buildTextYear()
Had a few errors in the new code.  First the main call to buildTextYear() was passing in the primary chrono_list rather than the chrono_tmp_list.  Second the isTextYear() method was returning the absolute coordinates rather than the relative, and it was including coordinates of end punctuation, which I fixed, so now the correct coordinates are being returned.