Added in period and repeating-interval subintervals for NthFromStart
Added in code to identify sub-intervals for period and repeating-interval.  The problem is it doesn't work very well.  It looks like we are missing some periods like "day".  This term is parsed, but it is never assigned to a period or sub-interval.  When adding in this parsing our performance goes down because it is not getting everything correct.  Need to work on this more and fine tune the parsing rules.