fixed parsing bug in hasYear()
I updated hasYear() to return the matching string instead of the original text.  If there are any punctuation marks at the end of the text it was returning these and giving an error.  Now I have it so that it operates on only the matching regex string, which does not include the punctuation.