Added special cases to buildPeriodInterval() and hasPeriodInterval()
Added a special cases to buildPeriodInterval() method:   it now looks for "tomorrow" and "today" specifically before going to the ML module to classify these as calendar intervals.

Added additional code to specificaly parse out the "daily" mention in "one-time daily" and similar phrases.  This includes weekly, monthly and yearly as well.